Moving AlphaVantage from "community" to "official integrations"
I am the cofounder of Alpha Vantage. My team (along with the repo creator @calvernaz) has validated the MCP integration in production environment and we have dedicated engineering resources to maintain the repo going forward. 

Given the production-ready status and maintenance commitment of the repo, can we move it from "community" to the "official integrations" section of the README? Thank you!